Add full device and most recent event/telemetry information to compliance output
